refactor(plugin): use textlint-plugin-{text, markdown}
use processor plugin and remove internal plugin
