+++ package.json
@@ -65,6 +65,8 @@
     "rc-loader": "^1.0.0",
     "textlint-formatter": "^1.1.1",
     "textlint-plugin-html": "^0.1.2",
+    "textlint-plugin-markdown": "^1.0.1",
+    "textlint-plugin-text": "^1.0.1",
     "traverse": "^0.6.6",
     "try-resolve": "^1.0.1",
     "txt-ast-traverse": "^1.2.0",
